Add files via upload
In this task, i worked with the titanic dataset. i loaded the data, checked the important columns, and made some graphs to understand the data better.
I created charts to show how many passengers survived and how the ages are distributed. these visuals help to see patterns in the data.
all the code, data file, and graphs are included in this submission.